+++ packages/react/src/SelectPanel/SelectPanel.stories.tsx
@@ -53,8 +53,6 @@ export const Default = () => {
     <>
       <h1>Multi Select Panel</h1>
       <SelectPanel
-        title="Select labels"
-        subtitle="Use labels to organize issues and pull requests"
         renderAnchor={({children, 'aria-labelledby': ariaLabelledBy, ...anchorProps}) => (
           <Button
             trailingAction={TriangleDownIcon}
@@ -65,15 +63,12 @@ export const Default = () => {
             {children ?? 'Select Labels'}
           </Button>
         )}
-        placeholderText="Filter labels"
         open={open}
         onOpenChange={setOpen}
         items={filteredItems}
         selected={selected}
         onSelectedChange={setSelected}
         onFilterChange={setFilter}
-        showItemDividers={true}
-        overlayProps={{width: 'small', height: 'xsmall'}}
       />
     </>
   )
